Added ZlibOptions, DeflateOptions, GZipOptions, and BrotliOptions to System.IO.Compression.
New Public Types:
- System.IO.Compression.ZlibCompressionLevel
- System.IO.Compression.BrotliCompressionLevel
- System.IO.Compression.ZlibCompressionStrategy
- System.IO.Compression.ZlibFlushCode
- System.IO.Compression.ZlibMemoryLevel
- System.IO.Compression.ZlibOptions
- System.IO.Compression.DeflateOptions
- System.IO.Compression.GZipOptions
- System.IO.Compression.BrotliOptions

This is the implementation of the zlib, deflate, gzip, and brotli option types.

As for the streams, an encoder / decoder being public in the future, as well as using them in the streams internally, should allow to have a new public constructor to them accepting the new options type as well which would simplify everything in them.

Partially Fixes: https://github.com/dotnet/runtime/issues/42820